Move trigram to client/lib
