+++ client/lib/trigram/index.js
